+++ crates/amaru-ledger/src/state/volatile/series.rs
@@ -70,6 +70,10 @@ impl VolatileState for VolatileSeries {
         self.aggregate.resolve_cc_member(credential)
     }
 
+    fn cc_members(&self) -> impl Iterator<Item = &StakeCredential> {
+        self.aggregate.cc_members()
+    }
+
     // ----------------------------------------------------------------------------------- Proposals
     type Proposal = Existence<ProposalKind>;
     fn resolve_proposal(&self, id: &ProposalId) -> Existence<ProposalKind> {